requirements.txt
Flask-SQLAlchemy
Flask
pymysql
python-dotenv
sqlalchemy
sqlalchemy_utils
marshmallow-sqlalchemy
cryptography
